feat: add collapsed menu prototype
